view для intro, greeting, rules, stats
