+++ js/screens/intro.js
@@ -1,20 +1,13 @@
 import {changeScreen} from '../util';
-import {createDomTemplate} from '../create-dom-element';
-import showGreetingScreen from './greeting';
+import showGreeting from './greeting';
+import IntroView from '../views/intro-view';
 
 export default () => {
-  const moduleContent = createDomTemplate(`
-    <section class="intro">
-      <button class="intro__asterisk asterisk" type="button"><span class="visually-hidden">Продолжить</span>*</button>
-      <p class="intro__motto"><sup>*</sup> Это не фото. Это рисунок маслом нидерландского художника-фотореалиста Tjalf Sparnaay.</p>
-    </section>
-  `);
+  const intro = new IntroView();
+  intro.onClickNext = () => {
+    changeScreen(showGreeting());
+  };
 
-  const asterisk = moduleContent.querySelector(`.intro__asterisk`);
-  asterisk.addEventListener(`click`, () => {
-    changeScreen(showGreetingScreen());
-  });
-
-  return moduleContent;
+  return intro.element;
 };
