+++ js/views/game-view.js
@@ -1,0 +1,3 @@
+import AbstractView from '../abstract-view';
+import showHeader from '../header';
+
