+++ js/views/rules-view.js
@@ -1,0 +1,53 @@
+import AbstractView from '../abstract-view';
+import showHeader from '../header';
+
+export default class RulesView extends AbstractView {
+  get template() {
+    return `
+    ${showHeader()}
+    <section class="rules">
+    <h2 class="rules__title">Правила</h2>
+    <ul class="rules__description">
+      <li>Угадай 10 раз для каждого изображения фото
+        <img class="rules__icon" src="img/icon-photo.png" width="32" height="31" alt="Фото"> или рисунок
+        <img class="rules__icon" src="img/icon-paint.png" width="32" height="31" alt="Рисунок"></li>
+      <li>Фотографиями или рисунками могут быть оба изображения.</li>
+      <li>На каждую попытку отводится 30 секунд.</li>
+      <li>Ошибиться можно не более 3 раз.</li>
+    </ul>
+    <p class="rules__ready">Готовы?</p>
+    <form class="rules__form">
+      <input class="rules__input" type="text" placeholder="Ваше Имя">
+      <button class="rules__button  continue" type="submit" disabled>Go!</button>
+    </form>
+    </section>
+    `;
+  }
+
+  bind() {
+    const buttonStartGame = this.element.querySelector(`.rules__button`);
+    const inputPlayerName = this.element.querySelector(`.rules__input`);
+    const buttonBack = this.element.querySelector(`.back`);
+
+    inputPlayerName.addEventListener(`input`, () => {
+      if (inputPlayerName.value.length) {
+        buttonStartGame.removeAttribute(`disabled`);
+      } else {
+        buttonStartGame.setAttribute(`disabled`, `disabled`);
+      }
+    });
+
+    buttonStartGame.addEventListener(`click`, () => {
+      this.onClickNext();
+    });
+
+    buttonBack.addEventListener(`click`, () => {
+      this.onCLickBack();
+    });
+  }
+
+  onClickNext() {}
+  onCLickBack() {}
+}
+
+
